check login status on profile drawer implementation
