+++ screens/ProductDetailsScreen/styles.js
@@ -70,9 +70,9 @@ export default StyleSheet.create({
 		width: Layout.moderateScale(30),
 		height: Layout.moderateScale(30),
 		borderRadius: Layout.moderateScale(15),
-		backgroundColor: Colors.iconsec,
+		backgroundColor: '#7f7f7f',
 		borderWidth: 1,
-		borderColor: Colors.iconsec,
+		borderColor: '#7f7f7f',
 		justifyContent: 'center',
 		alignItems: 'center',
 	},
